Current State Node: Preserve original message with option to override payload
Modify this file along with api_current-state.html

This is a clone of PR 55 of the orignial node-red-contrib-home-assistant from declankenny.

This change adds a checkbox to the UI, override_payload to the Current State Node, allowing the user to preserve the incoming message payload or override them with the entity.id and status - replicating the existing behavior.

Control defaults to selected in the UI form to replicate existing design.

On the back-end, the code will trap for undefined values coming from this and default to selected. This ensures backward compatibility for existing nodes.